Web Component JSX
Further expanded my own JSX factory setup! Did some testing with how React handles typing for JSX elements, and used that to make mine more consistent with that.

I noticed that all JSX components are of the type `JSX.Element` in regular JSX, even with a more specific tag being passed in. So I realized that `HTMLElement` would probably be the fitting type for the vanilla version here.

When I tried to use SVG elements though, I found out that they are under a different tag map interface, so I had to account for that inside of the factory definitions.

https://stackoverflow.com/questions/41557309/typescript-jsx-without-react
https://www.typescriptlang.org/docs/handbook/jsx.html

I haven't gotten Component types working yet. I'm not totally sure how all of this should work quite yet, so I'm learning little by little what the best way to set it up should be.

With this update, I noticed that the properties on the elements are being properly type-checked, which is awesome! But it makes it hard for some properties, as right now I'm getting the types from the element classes themselves, which means for attributes like `part`, it doesn't like when you assign a string to it, because it's of the `DOMTokenList` type, which uses methods to be interacted with instead.

So I think Texsaur was smart in using separate definitions for the element attributes themselves, which makes it behave more like you're working with HTML elements and their direct attributes, rather than their properties, which are more complex.

Yeah, I'm suprised how semi-nice this has worked so far. I didn't expect you to be able to customize the behind the scenes implementation of JSX like this! It's not quite too straightforward right away though. I'm suprised there haven't been more projects to try and do it like this, it seems like a fairly nice dev setup to structure it like this. It is a bit quirky though. Maybe because I'm trying to use it for something other than React? Haha.

I am curious about trying to use JSX for NBT though, that sounds kind of interesting. It sounds very neat, and very ugly at the same time XD

I was having some trouble with `git commit --amend --no-edit --date=""` today, I can't get it to work for the life of me! Really annoying. Not sure what's going wrong with it. I put in multiple date formats with the time and date that I want to use, and they all either error out during the Rebase, or they put a super wonky date on the commit instead. My old scripts that I've been using for a while don't seem to work anymore. Aggh! If you have any info about it, please let me know. It seems to have been more broken in the last year or so. A newer update to Git maybe? It used to only not work on my Chromebook, so I would go to my Mac. Now my Mac is having the same weird issues on there too :(

When in doubt, I'ma keep coding :)