JSXImportSource
Nearly got this working! It's fine at the TS level, but Vite isn't liking it completely yet. Found out I can get the module resolution working with a `package.json` exports alias, which is how React itself sets this handling up. The problem with Vite at the moment is locating `jsxDEV` in the live server, and `jsx` in production. I don't fully get how the module references to those are supposed to work yet.

https://www.typescriptlang.org/tsconfig#jsxImportSource