Prototype Analyzer
Been working crazily on this the last day or so! I was having some good trouble with figuring out how to do this. Now that I have it, it works really well :)

The concept is that you can get the prototype chain of a given class constructor object, simply by going into the prototype of the class recursively, until you get down to the plain `Object` object's prototype, which is prototyped with `null`.

Now I can use this to automatically determine which properties in the prototype are mutable, and which are readonly, hence being only accessible with setters. So you could also check whether a specific property is readonly, the other version of that. This is very neat!