Function Components!
I need to heavily improve the types for `jsx()`. I wish JSX was handled better at the type level in TS, it's a lot more magic than I wish it were. Things based on the JSX namespace is a bit messy because it doesn't validate the return types of JSX calls, nor does it throw errors for undefined JSX namespace type implementations. There's also disparity between the types of the `jsx()` function pragma implementation, versus when it's called with the XML syntax.