Async Components
This is probably a bad idea, but I've wanted this concept for a while.

I want to be able to use async await with JSX, it doesn't seem to be effective for some reason, so I'm trying to see what it looks like if it were to be done here.

I think the hard part is using JSX as templating, but getting proper type checking for it, and also dealing with Promises when using the components themselves.

I think I also need to find a trade-off between using class components/Web Components, versus function components like React. I like functions for templating, but classes for stateful components. I also think async functions for long-running tasks are great for HTML handling too. I just wish TS would have better type inference for JSX's types. :(

It also seems bad that this inherhently marks everything as async, even though it doesn't have to be. I'm curious how I should/could go about handling that.