Async API Fetching
Now I'm seeing why async JSX isn't quite logical. It makes sense that an implementation function can handle the fetching instead, and then after that JSX can render that fetched data. I think the point is to abstract the network handling *out* of your component, so it isn't handled there specifically. Or rather, it shouldn't be in the root scope of the component, because that will block the rendering of the JSX tree in that way. Using something like `<Suspense/>` makes sense, in that the component will handle when rendering certain things should happen. The JSX engine shouldn't need to wait for async state/props operations, that should be done by state handling and or outside of the JSX templating step.

I'm glad I tried this out, because now it makes sense as to why it's not ergonomic. I think this works amazingly for building regular element components using standard functions and async functions though, using complete JSX (DOM) trees with async data feels like a really natural model to me. One example of that is to fetch the data accordingly, then build the JSX tree in it's own component function using that pre-fetched data (in terms of what JSX is concerned). It doesn't need to know that the data was async initially, you handle that before trying to use it with JSX.

I have to say, it would be great to do this `Promise.all()` stuff with JSX templating/mapping though, I guess it also makes sense that that part can be done outside of JSX though, and moved into it after it completes. I think the point of being able to do that probably should be done with state instead maybe, since that pattern appears to not be as feasible as I'd thought.