Async Props Dependency Injection
That's it! Async handling probably should be managed with dependency injection, rather than have it be described from within the JSX itself. At least I like that method a little more I feel.

Also realized I accidentally was missing a config option for the Vite configuration, as well as an additional pragma for JSX in production. I'm not sure what `jsxs()` is specifically meant for, but aliasing it in the implementation's exports seemed to fix it for now at least :p